Add Big O Tests
This tests to see if the time it takes to perform
`Base#self.render` with increasing number of objects, will increase
the time to complete in a linear fashion.

Each test will run the following number of objects:
1, 10, 100, 1000, 10000.

We have two tests, one for basic objects, and one for active record
objects.