Add Yard gem for documentation
It is added in development runtime as it is only to be used in
development.

Official website of yard: https://yardoc.org/

yard github: https://github.com/lsegal/yard

We will eventually use http://rubydoc.info to parse the yard tags.

In order to view docs in development, do execute `yard server`,
then navigate to localhost:8808