Ensure view works with unit tests.
Put the defaults for a View into keyword arg defaults. I think this gives
us a lot more flexibility in various possible use cases as well as in
testing.

I caught a bug in 'def <<' in which we were keying the early escape off of
whether or nor the field was a key in the @fields hash, where we should
have been looking for field.name.

Lastly, I removed the unnecessary instance variables since they were
already attr_readers.