Support Hash by using AutoSerializer as default
AutoSerializer will delegate between HashSerializer to serialize
Hash objects or to PublicSendSerializer to serialize objects with
dot notation attributes.

We also don't iterate over all objects that respond to `:map` anymore
because Hash responds to `:map`, but we wouldn't want that. So we
have to be more specific on what objects we treat like an array.

Add spec contexts for Hash objects.