Remove Optimizer
Ok, Optimizer is not working and is causing multiple problems.
Until we come up with a better solution, I think it is best to remove
Optimizer so that we can start accepting ActiveRecord objects that
have relations with other objects.

We should re-introduce Optimizer at a later time when we have a working
solution.