Introduce View::Manager, Composer, Builder
Instead of having `Base` know about `FieldMap` directly, we only expose
`View::Manager` to the `Base` class. When the `Base` class needs to
add/exclude fields to a view, or to add a view, the `Base` class will
use the `View::Manager`'s APIs.

If it is a view change, `View::Manager` is then responsible for
delegating work to `View::Builder`.

If it is a `Field` addition or exclusion from a view, it will delegate
the work to `View::FieldMap`.

`View::Manager` will also tell the `View::Composer` to `compose` the
fields based on the `View::Builder`'s ancestry.