Remove rails app from spec directory
Instead of depending on a full rails app to setup a database, migrations
and models, configure ActiveRecord in spec/activerecord_helper.rb using
an sqlite3 in memory database.

By removing the rails app, autoloading was disabled when running
specs/benchmarks. This uncovered some missing requires on both lib/ and
spec/.