Localize private flag to class and not module
This private flag was accidentally placed that the top level module,
causing many methods and classes set at the top level to be considered
private when it shouldn't have been.

Take a look at https://www.rubydoc.info/gems/blueprinter/Blueprinter
it has the `configure` method as private, when it should be public.

Moving this private flag down to the Extractor class level should
fix this.