docs: update field-level conditional settings
The docs were not updated to reflect that both if and unless now
expect three arguments for the conditional proc, even in the field settings.
Passing two arguments as the previous version of the docs propose will yield
a deprecation notice. This commit updates the docs so that it indicates
how to use the three version argument.