feature: conditional view rendering
Prior to this change, applying conditions to a view required annotating
all fields individually.

This change allows views to be defined as `view :name, if: condition`,
using the same `->(field_name, object, options) { ... }` callbacks as
used on fields. These callbacks are propagated to the fields, and are in
addition to (not in replacement of) any checks defined on the fields.

Tests are added to cover the logic of `unless` and `if`, and
`Field#skip?` is altered to allow a field to have both `unless` and `if`
defined.

Defining a view-level condition on an child blueprint does not affect
the parent view of the same name - this behavior may be unexpected, and
it is advised not to add conditions to inherited views, since this may
represent a security flaw.

Signed-off-by: Alex Kalderimis <alex.kalderimis@gmail.com>